refactor: move get_absurd_backends to backends.py (co-locate backend-lookup helpers)
