+++ tests/test_results.py
@@ -9,8 +9,9 @@ from django.db import connections, transaction
 from django.tasks import TaskResultStatus
 from django.tasks.exceptions import TaskResultDoesNotExist
 
+from django_absurd.backends import get_absurd_backends
 from django_absurd.params import AbsurdSpawnParams
-from django_absurd.queues import get_absurd_backends, get_absurd_client
+from django_absurd.queues import get_absurd_client
 from tests.models import Payload
 from tests.tasks import add, boom, echo
